+++ e3nn_jax/__init__.py
@@ -58,6 +58,7 @@ from e3nn_jax._src.dropout import Dropout
 from e3nn_jax._src.mlp import MultiLayerPerceptron
 from e3nn_jax._src.graph_util import index_add, radius_graph
 from e3nn_jax._src.reduced_tensor_product import reduced_tensor_product_basis, reduced_symmetric_tensor_product_basis
+from e3nn_jax._src.symmetric_tensor_product import SymmetricTensorProduct
 
 __all__ = [
     "config",  # not in docs
@@ -133,4 +134,5 @@ __all__ = [
     "poly_envelope",
     "reduced_tensor_product_basis",
     "reduced_symmetric_tensor_product_basis",
+    "SymmetricTensorProduct",
 ]