+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `e3nn.SymmetricTensorProduct` operation: a parameterized version of `x + x^2 + x^3 + ...`.
 
 ## [0.10.0] - 2022-10-05
 ### Added